tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-timing-function_ease-in-out.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-timing-function: ease-in-out</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animation-timing-function_tag">
   <meta name="flags" content="dom">
   <meta name="assert" content="ease-in-out is a valid
   animation-timing-function">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes move {
         from {
            left: 0px;
         }
         to {
            left: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
      }
      .test {
         animation-name: move;
         animation-duration: 2s;
         animation-fill-mode: forwards;
         animation-timing-function: ease-in-out;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         function aboutEqual(observed, actual) {
            var maxDeviation = 5;
            return (Math.abs(observed - actual) < maxDeviation);
         }
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 500);
         div.addEventListener('animationstart', function() {
            clearTimeout(timeout);
            setTimeout(function() {
               var observed = getComputedStyle(div, null).left;
               var passed = aboutEqual(parseFloat(observed), 19);
               test(function(){assert_true(passed);}, testName); done();
               log.data = (passed ? 'PASS if the animation occurred.'
                  : 'FAIL (got ' + observed + ')');
            }, 600);
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The square below should move right, accelerating at the start and
   decelerating at the end.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
